Fix audit, dockerhooks and distros patches cross compile and CI
