+++ daemon/dockerhooks/dockerhooks_windows.go
@@ -1,12 +1,10 @@
 package dockerhooks
 
-import (
-	"github.com/opencontainers/specs/specs-go"
-)
+import "github.com/opencontainers/specs/specs-go"
 
 // Generate docker hooks Prestart and poststart specs
 func Generate(specHooks specs.Hooks, configPath string) specs.Hooks {
-	return nil
+	return specHooks
 }
 
 // TotalHooks returns the number of hooks to be used
